mocks: unify setPreimageAttached in lookup.ts, drop _-prefixed internal exports
Addresses Task 8 code review (issues #1 and #2, plus #5):

- Move setPreimageAttached / clearPreimageAttachments into
  sdk-ecalli-mocks/src/general/lookup.ts where the attached-preimage map
  actually lives. The ecalli being mocked is `lookup` (general), not
  `query`/`solicit`/`forget` (accumulate) — the new placement matches
  that boundary and eliminates the cross-directory import.
- Drop the `_setAttached` / `_clearAttached` underscore-prefixed internal
  exports. Hex encoding now happens in exactly one place.
- Restore the length-validation throw that the plan's snippet had and
  the first implementation pass dropped — setPreimageAttached now
  throws if the hash blob isn't 32 bytes.
- Skip the readBytes+hex work on the hot path when the attached map is
  empty (most lookups in tests that don't use setPreimageAttached).
- resetPreimages still clears the attached map — via the public
  clearPreimageAttachments() — so resetAccumulate() without a
  full resetAll() still leaves no cross-test leakage.

Top-level re-exports moved from the accumulate section to the
general/lookup section to reflect the new home.
